requirements.txt
aiogram==2.25.1
easyocr==1.7.1
opencv-python==4.8.1.78
python-doctr==0.7.0
psycopg2==2.9.3
sqlalchemy==2.0.21
ultralytics==8.0.81
